Исправил интрефес журнала и страницу редактирования студентов

+++ src/main/java/ru/innopolis/stc9/earth_stc9/services/JournalService.java
@@ -3,9 +3,7 @@ package ru.innopolis.stc9.earth_stc9.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.innopolis.stc9.earth_stc9.db.dao.JournalDAO;
-import ru.innopolis.stc9.earth_stc9.pojo.Journal;
-import ru.innopolis.stc9.earth_stc9.pojo.Subject;
-import ru.innopolis.stc9.earth_stc9.pojo.UserForJournal;
+import ru.innopolis.stc9.earth_stc9.pojo.*;
 
 import java.util.List;
 
@@ -24,27 +22,43 @@ public class JournalService implements IJournalService {
 
     @Override
     public boolean addJournal(Journal journal) {
-
+        if (journal == null) {
+            return false;
+        }
         return journalDAO.addJournal(journal);
     }
 
     @Override
     public boolean updateJournal(Journal journal) {
-
+        if (journal == null)
+            return false;
         return journalDAO.updateJournal(journal);
     }
 
     @Override
     public boolean deleteJournal(int id) {
-
+        if (id == 0)
+            return false;
         return journalDAO.deleteJournal(id);
     }
 
     @Override
     public List<Subject> getSubjectforJournal(String nameGroup) {
-
         return journalDAO.getSubjectGroup(nameGroup);
     }
 
+    @Override
+    public List<User> getStudentsFromGroup(String groupname) {
+        return journalDAO.getStudentsFromGroup(groupname);
+    }
 
+    @Override
+    public List<Lesson> getThemeFromSubject(String subjectname) {
+        return journalDAO.getThemeFromSubject(subjectname);
+    }
+
+    @Override
+    public UserForJournal getEntryFromJournal(int id) {
+        return journalDAO.getEntryFromJournal(id);
+    }
 }
